requirement.txt
altgraph==0.17.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
packaging==23.2
pdfminer.six==20231228
pefile==2023.2.7
pycparser==2.21
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.12
pywin32-ctypes==0.2.2
